+++ lib/hnco/functions/cancellation.hh
@@ -51,7 +51,7 @@ namespace function {
   protected:
 
     /// Bit vector size
-    size_t _bv_size;
+    int _bv_size;
 
     /// Buffer
     std::vector<double> _buffer;
@@ -77,7 +77,7 @@ namespace function {
     }
 
     /// Get bit vector size
-    size_t get_bv_size() { return _bv_size; }
+    int get_bv_size() { return _bv_size; }
 
     /// Evaluate a bit vector
     double eval(const bit_vector_t& x);
