+++ lib/hnco/functions/decorators/function-controller.hh
@@ -51,7 +51,7 @@ namespace function {
     ///@{
 
     /// Get bit vector size
-    size_t get_bv_size() { return _function->get_bv_size(); }
+    int get_bv_size() { return _function->get_bv_size(); }
 
     /// Get the global maximum
     double get_maximum() { return _function->get_maximum(); }
