+++ lib/hnco/functions/decorators/function-modifier.hh
@@ -70,7 +70,7 @@ namespace function {
     ///@{
 
     /// Get bit vector size
-    size_t get_bv_size() { return _function->get_bv_size(); }
+    int get_bv_size() { return _function->get_bv_size(); }
 
     /** Get the global maximum.
         \throw Error */
@@ -135,7 +135,7 @@ namespace function {
     ///@{
 
     /// Get bit vector size
-    size_t get_bv_size() { return _map->get_input_size(); }
+    int get_bv_size() { return _map->get_input_size(); }
 
     /** Get the global maximum.
         \throw Error */
@@ -192,7 +192,7 @@ namespace function {
     ///@{
 
     /// Get bit vector size
-    size_t get_bv_size() { return _function->get_bv_size(); }
+    int get_bv_size() { return _function->get_bv_size(); }
 
     /** Get the global maximum.
         \throw Error */
