+++ lib/hnco/functions/equal-products.hh
@@ -116,7 +116,7 @@ namespace function {
 
 
     /// Get bit vector size
-    size_t get_bv_size() { return _numbers.size(); }
+    int get_bv_size() { return _numbers.size(); }
 
     /// Evaluate a bit vector
     double eval(const bit_vector_t&);
