+++ lib/hnco/functions/factorization.hh
@@ -56,7 +56,7 @@ namespace function {
     size_t _second_factor_size;
 
     /// Bit vector size
-    size_t _bv_size;
+    int _bv_size;
 
     /// Init GMP data structures
     void init();
@@ -97,7 +97,7 @@ namespace function {
     void load(std::istream& stream);
 
     /// Get bit vector size
-    size_t get_bv_size() { return _bv_size; }
+    int get_bv_size() { return _bv_size; }
 
     /// Evaluate a bit vector
     double eval(const bit_vector_t&);
