+++ lib/hnco/functions/jump.hh
@@ -42,7 +42,7 @@ namespace function {
     public Function {
 
     /// Bit vector size
-    size_t _bv_size;
+    int _bv_size;
 
     /// Gap
     int _gap;
@@ -53,7 +53,7 @@ namespace function {
     Jump(int bv_size, int gap);
 
     /// Get bit vector size
-    size_t get_bv_size() { return _bv_size; }
+    int get_bv_size() { return _bv_size; }
 
     /// Evaluate a bit vector
     double eval(const bit_vector_t&);
@@ -86,7 +86,7 @@ namespace function {
     public Function {
 
     /// Bit vector size
-    size_t _bv_size;
+    int _bv_size;
 
     /// Gap
     int _gap;
@@ -97,7 +97,7 @@ namespace function {
     DeceptiveJump(int bv_size, int gap);
 
     /// Get bit vector size
-    size_t get_bv_size() { return _bv_size; }
+    int get_bv_size() { return _bv_size; }
 
     /// Evaluate a bit vector
     double eval(const bit_vector_t&);
