+++ lib/hnco/functions/long-path.hh
@@ -63,7 +63,7 @@ namespace function {
     public Function {
 
     /// Bit vector size
-    size_t _bv_size;
+    int _bv_size;
 
     /// Prefix length
     int _prefix_length;
@@ -81,7 +81,7 @@ namespace function {
     ///@{
 
     /// Get bit vector size
-    size_t get_bv_size() { return _bv_size; }
+    int get_bv_size() { return _bv_size; }
 
     /** Check for a known maximum.
